Implement opentelemetry for the backend
This change just follows the instructions on the GCP docs just to get it working
- https://cloud.google.com/run/docs/tutorials/custom-metrics-opentelemetry-sidecar
- https://cloud.google.com/stackdriver/docs/instrumentation/setup/go
- https://github.com/GoogleCloudPlatform/golang-samples/tree/main/opentelemetry/instrumentation/app

I did not add any spans yet. This is only to get the functionality working.

Additionally, I converted all slog entries to use their xxContext variant which automatically
associates the log entry with the given trace.

Future things we can do:
- Return a request id so that users can report it and we can easily search our logs
- Add more detailed spans

Other cool things that we get out of this:
- The spanner client is instrumented with opentelemetry already. That means we get some useful spans.

Change-Id: Ifb856db39aedc2297f2ee2ae23cc023f2c3f182f