Fix overview page error propagation
Currently, we use Lit's async task to retrieve a feature page
from the backend. While the rest of the callbacks work, the error
callback does not work. Also, we pass the task deep into the tree.

Instead of doing that, I introduced a new TaskTracker interface.
It is a generic that takes in the expected type and the type of Error.

This TaskTracker type is now the entity passed in not the task. We hook
use the onComplete and onError callbacks of the task to set the value of
the immutable TaskTracker object. Which will re-render correctly.

Now we also handle 400 requests from the overview page. It shows
a different message too.

Other fixes:
- The error from openapi comes as a instance of type "object".
  The code has been corrected to account for that. Kept the error check
  just in case.
- Move the errors for the api into their own file.
- Create a constants file for the links

Future improvements:
- Doing client side form validation and showing an error below the form field

Change-Id: Ib2f881db0fa9c6a5e15531f3d911ba2bc9eec4cc