Allow progressive loading of chart and other chart UX fixes (#251)
* Allow progressive loading of chart and other chart UX fixes

I modified getFeatureStatsByBrowserAndChannel to be an async generator. With this,
we can generate the data table with the current data we have and redraw the chart
as it continues to load. This will give the illusion / better UX that the data is loading
while the user waits.

While it waits for the data to load, we know we are waiting because of the
task status that is read. This status is passed into gchart component to
place a site wide overlay to prevent the UI from being touched while loading.

I also made a common startMetricsTask method that can be used. This makes sure
that the status always comes from the task.

Bug fix:
- There was one bug where a user loads a really small window of data, they would
  get the data back before the gchart library is finished loading. I buffer that now
  in the event that happens.

Future modifications:
- The data table is constantly rebuilt which is good to make sure we aren't missing data.
  However, we could probably do some smart appending. But that can be done later on.

Change-Id: I2ee351170ff08e5bb88bbfd19e230c78b00f936b

* move overlay to own file

Change-Id: I6bd437c5465bab7a400c390fd38bf7669cf75472

* touchups

Change-Id: Id195580fd0cd044dc9c4e2131401fd44208c25d2

* final touches

Change-Id: Ice843d9539102be7a9b1d1f033ad03e82942fe9e