Create reusable base devcontainer images (#335)
This is the beginning of the work described in https://github.com/GoogleChrome/webstatus.dev/issues/329

It will cover the first two check boxes on that issue.

This change creates a new GitHub CI workflow that will refresh the images:
- Every tuesday (the dependabot notices will be configured for Monday)
- Ad-hoc runs that can be triggered on the Actions page

Two of the images are now used in the existing build GitHub action.

We have to split them up because the GitHub runner runs out of space. We could clean up more
and try to fit it all into one image but that can come in the future.

For now, this should reduce the time spent in Github Action machine hours.

Change-Id: Ic44aad153de2f45246821220707e2e28b717ffbb