Fix opentelemetry config (#499)
Opentelemetry was recently updated in #476

I did not read the Changelog but there's an important notice there from the release [notes](https://github.com/open-telemetry/opentelemetry-collector-contrib/releases/tag/v0.104.0):

> Expansion of BASH-style environment variables, such as $FOO is no longer supported by default. If you depend on this syntax, disable the confmap.unifyEnvVarExpansion feature gate, but know that the feature will be removed in the future in favor of ${env:FOO}.

I started to see problems in staging when I deployed it. This change fixes it to be consistent with other environment variables.

Change-Id: I5a738594cebf7e62a78c3ddd99c1dc799b0ef853