Fix gzip compression in GCP
I fixed GZIP compression. However, when I deployed it to GCP,
it still did not work. Turns out that nginx turns off gzip compression
when there's a Via header sent. (which is what GCP load balancers do) [1]

This change follows the recommendations from the GCP docs [1] and modifies the
test to use a Via header to simulate being in a GCP environment to be close
to what is used live.

[1] https://cloud.google.com/load-balancing/docs/https/troubleshooting-ext-https-lbs#compression-not-working

Change-Id: I9eb3c4ae224eb5f585655afb75bd21abc6743050