chore: Support new structure of web features data.json
The Web Features repo changed the structure of the data.json in this PR: https://github.com/web-platform-dx/web-features/pull/1060

Instead of being a map of features, it is now an object with three top level keys:
- features
- groups
- snapshots

This is a breaking change and has caused us to start ingesting features as if they had IDs: features, groups, snapshots

This commit fixes that by updating the defs.schema.json to be the latest version (v0.10.0). This file comes from the web features repo.

After auto-generating the code for that schema file, this commit adjusts the code to use it.

We previously had v0.6.0 of the schema. And since then, some optional fields have become required fields. So you will see the
removal of some nil checks (since now the field is not nullable).

Change-Id: I9861421bc53ec8966eee80c33742d2a7e94d744c