Add post process step to the web features parsing
We need to remove the range character from some of the fields.

This post process step rebuilds each of the structs. We do this so that
go-exhaustruct will let us know if any new fields are added. And at that time
we can look how to handle it.

The alternative would be to have conditionals for the specific fields that we know of.
But during regeneration, we would not get alerted if a new field was added.