Refactor: introduce ServiceElement & WebstatusServicesContainer (#547)
This change introduces the notion of ServiceElements. These are LitElements
that generally do not render anything except <slot></slot>. They need early initialization
and provide data via @lit/context.

There is now a new ServiceElement that renders only that. That lives inside the services directory.

I looked at the components that only render `<slot></slot>` and moved them into the services directory and changed
them from extending `LitElement`` to this new `ServiceElement`. Those components are:
- webstatus-app-settings-service
- webstatus-gcharts-loader-service
- webstatus-api-client-service
  - This was pulled from webstatus-app-settings-service because we will get rid of webstatus-app-settings-service after moving to firebase auth and creating a dedicated service just for auth.

All of these service elements are rendered in a new component called WebstatusServicesContainer and
that wraps the UI components so it can pass the data down via context as needed.

Why accept this change:
- I was about to add more "service"-like elements for authentication and firebase. This centralizes all of the services inside the container.
- Improved Code Organization: Clear separation of concerns between UI and service logic.
- Future-Proofing: The new ServiceElement class and WebstatusServicesContainer provide a solid foundation for introducing new services as the application evolves (e.g. Firebase App and Auth)
- Increased Testability: The clear structure and decoupling of services facilitate the creation of more focused and robust unit tests.