Add services for firebase/app & firebase/auth
This change adds two new services called `webstatus-firebase-app-service` and `webstatus-firebase-auth-service`.

These two services are responsible for handling the initialization of firebase app and firebase auth.

The settings for these two services are passed in as environment variables like other settings in index.html. The final index.html that is served by nginx fills it in one time before responding with index.html

webstatus-firebase-app-service:
- If the environment variables are not empty, it will create a firebase app and provide it via context.

webstatus-firebase-auth-service:
- Consumes the firebase app (if it was indeed created) and creates a firebase auth config.
- The firebase auth config is configured to do github authentication
- Also, sets up the user callback which will detect if a user is signed in or signed out. The status of the user is provided as a separate context that UI components can read later on.

This does not render the button yet. That will come as a separate change.