Return browser version for feature implementation
This change adds a new optional string field to the
BrowserImplementation component for the browser version that the feature
became available.

The change also modifies the database query to return this data too

This completes the backend part of #306.

There still needs to be a frontend implementation to complete that issue.