Helper to get credentials while on GCP infra
This helper code is taken directly from
https://cloud.google.com/docs/authentication/get-id-token#metadata-server

It will be needed by the uma export service that will come in a future PR
(you can see it in action in #616). This is similar to the fetch_id_token
that happens on [ChromeStatus](https://github.com/GoogleChrome/chromium-dashboard/blob/8fd512af666290e13173176268b9a4e28bbc9c63/internals/fetchmetrics.py#L53)

FindCredentials is a method to get credentials from the metadata-server (found only when on GCP infra).
As a result, there is a skipped test that will need to be run directly on GCP infra. (helpful if we move to GCP Cloud Build)