Add generic entity reader (#666)
Currently, we have generic methods for writing rows to the spanner database
but not for reading. This adds a new entityReader type. Additionally,
since some of the writers already do reading, I made the writers inherit
from this new reader.

Future PRs will use this as part of the split up of #616.