Refactor: Centralize worker start logic
Previously, each process had to create their own channel
and send the jobs into it. This led to some near duplicate code
for each process. In anticipation of two new processes (1. Chromium Enum & 2. UMA Export),
I refactored the workerpool logic to include the creation of the job channel and feeding the jobs into that
channel.

This change helps out with the two existing processes (BCD and WPT).