Add main entrypoints for the uma and chromium jobs (#681)
In go, an application needs to have a main package so it knows where to start.
This changes adds a main package for each the uma and chromium jobs.

Also, add a command to run the chromium job locally. (We can't run the uma job locally because locally we don't have access to UMA Export)

Part of the #616 split up.

Alot of these changes are from the go module file update which happens automatically.