Make firebase settings optional when deploying
While waiting on getting the GitHub app setup, we don't want prod to show
a login button. This makes those fields optional. Then we change the prod configuration to temporarily remove those fields.

Without this, a login button will appear in prod but it won't function (since the prod GitHub
app is not setup).