Use Target Architecture Argument for docker build (#726)
In the spanner emulator, a version of wrench is downloaded. But it is hardcoded to amd64.

This change uses the auto populated TARGETARCH argument.

Now, we know which architecture to use when downloading the wrench CLI. Should fix the problem for those running the repo on a Macbook.

See: https://github.com/GoogleChrome/webstatus.dev/issues/720#issuecomment-2394365939