Add variables to tune scaling for backend & frontend (#754)
By default cloud run will scale down to zero. This
change exposes variables for the minimum and maximum
number of instances.

For prod, we set the min to 1. For staging, we set the min to 0 (for now)

For prod and staging, we set the max to 100. This is the default anyhow.