Add PrecalculateBrowserFeatureSupportEvents method
Depends on: #865

This method is the entrypoint to calculate the events stored in the BrowserFeatureSupportEvents table introduced in #865.

It reads all of the web features, browser releases and feature availabilities. With that, it builds the event log.

In order to read all that data, small helper methods were created to assist.

This PrecalculateBrowserFeatureSupportEvents method does not take in any input and just uses existing table data.

Fixes #834