Web Features workflow now calls Precalculate method for feature support events (#881)
This now mandates that the WebFeatureSpannerClient implements
PrecalculateBrowserFeatureSupportEvents (which already happened in #872). Since the
spanner client already implements this in that mentioned PR, we don't
have to do anything else besides modify the expected interface.

Also, modified the tests to accomodate for the fact that we are calling that now.

Addresses: #836