Change feature_lag_count API to missing_one_implementation_counts (#879)
* Change feature_lag_count API to missing_one_implementation_counts

This will match the name of the milestone we are tracking.

For example: /v1/stats/features/browsers/fooBrowser/missing_one_implementation_counts?browser=barBrowser&browser=bazBrowser

In this instance, we want to get the time series data for fooBrowser where it is the only browser missing an implementation compared to barBrowser and bazBrowser.

The actual implementation where it calls the database will occur in a future PR.

Other changes:
- Change the the browsers and featureIDs query parameters to singular. These are not used yet. So it's not a breaking change yet. Additionally, fix the comment that stated these were comma separated. They are not. We use the default query parameter settings documented [here](https://swagger.io/docs/specification/v3_0/serialization/#query-parameters). As a result, to specify an array, you supply the parameter multiple times.
- Created a common SupportedBrowsers schema in openapi.

* address feedback