Pass timeout config to GCP workflow
Currently: When the GCP workflow calls the cloud run job, the cloud run
job runs with a certain timeout configured. But the workflow uses the default timeout.

This leads to the workflow being marked as failed within the default time. But
the job actually is still running.

This change makes it so the workflow uses the same timeout as the job.