Move web features workflow from service to job (#888)
* Move web features workflow from service to job

Background:

The web-features workflow was the first workflow and it was built as a
http service (similar to cron jobs in other projects). However, the GCP
workflow engine only allows a http timeout of 30 mins. The new precalculate
method puts us over that limit as it adds many rows on the first run (>1M rows).
While we could create another workflow just for that precalculate method,
that's just another piece of infrastructure to manage. We have #474 in
the backlog. And this demand for a higher timeout means we could just
address that issue.

The rest of the workflows in use are not http services and are instead
cloud run jobs. And cloud run jobs allows a much higher timeout of 1 day.

Change:
- Fixes all the points in #474
- Uses the same reusable single_stage_go_workflow terraform module as the other workflows.
- Sets the job to have a timeout of 2 hrs

Outstanding things to be fixed in future PR(s)
- As-is, local workflow won't work because of the sheer number of rows. Will allow of subset of the data to be consumed locally by specifying a time range.

* Add batchWriter interface for spanner (#900)

The spanner emulator does not currently support the batch write API. Probably
because it is in Preview on GCP. But we should leverage this here for our workflows.

This new interface is used by the Client. Upon construction, it determines
which batch writer to use.