Warn for certain DailyChromiumHistogramMetrics errors
Previously, we returned an error for any error when ingesting the metrics.

There are certain errors we should warn instead of stopping execution.

Added DailyChromiumHistogramMetrics specific errors and modified the logic
to check for a subset of the errors. That subset will warn instead of returning
the error back up the stack.

Other changes:
- Use ErrQueryReturnedNoResults in places it should be to be consistent with other places. ([example](https://github.com/GoogleChrome/webstatus.dev/blob/2901887e4c0201aa40c0a2304c24669c590f026a/lib/gcpspanner/client.go#L378-L380))