Add BaseChartsPage common component (#1057)
* Add BaseChartsPage common component

BaseChartsPage is a component that all pages with charts can use. It
provides a renderDateRangePicker() method that any child class can use
in its main render method. It also centralizes the event handling for
the start and end date changes.

We also rename the existing updateFeaturePageUrl util method to updatePageUrl.
The BaseChartsPage component uses it update the page's url appropriately
when the date range changes.

* Apply suggestions from code review

Co-authored-by: Daniel Smith <56164590+DanielRyanSmith@users.noreply.github.com>

---------

Co-authored-by: Daniel Smith <56164590+DanielRyanSmith@users.noreply.github.com>