Return 400 Bad Request for Invalid Page Tokens (#1080)
Background: This change is in response to some 500 errors I saw this week.

Currently, if a user passes an invalid page token, they will receive a 500 error. This is because the page token format is a Base64 encoded string that contains a format potentially unique to each endpoint, and its validity cannot be determined until the database tries to parse it.

This change improves the user experience by returning a more appropriate 400 Bad Request error for invalid page tokens.

This change introduces a new error type, ErrInvalidPageToken, in the adapter layer.

For the spanner adapter, when it receives the invalid cursor error, it will raise the ErrInvalidPageToken error. The server will then return a 400 error to the client.

This pattern is adopted for all endpoints that use page tokens.