Factor in ExcludedFeatureKeys for aggregation calls (#1129)
Unlike the feature specific statistic spanner calls, the aggregation spanner
calls account for all features. However, it does not account for the ExcludedFeatureKeys table.

That table contains manually entered in human readable feature keys that should be excluded.

This change gets the spanner UUID for each of the feature keys and then uses it in the aggregation calls
to ignore those features when doing the calculation.