Remove the max line from the global support chart
We are currently at 1k+ features, so I would expect the max line to be
there. However, we currently calculate the max to be the max among the
current values. We should remove that line for now as it may confuse.

Instead, we need the timestamp of a feature's inception.

We should revisit once https://github.com/web-platform-dx/web-features/issues/714
is tackled.