feat: Add ListBaselineStatusCounts to retrieve cumulative baseline feature counts in spanner (#1132)
* feat: Add ListBaselineStatusCount to retrieve cumulative baseline feature counts

This PR introduces the `ListBaselineStatusCount` function and its associated helper functions to retrieve a cumulative count of baseline features over time.

The `ListBaselineStatusCount` function works similarly to `ListBrowserFeatureCountMetric` by:

1. Calculating an initial cumulative count of baseline features up to the specified `startAt` date.
2. Retrieving subsequent baseline feature counts within the specified date range (`startAt` to `endAt`).
3. Accumulating these counts to provide a cumulative view of baseline feature adoption over time.

Currently, the function only supports retrieving counts based on the `LowDate` field in the `FeatureBaselineStatus` table. However, support for `HighDate` can be easily added in the future by expanding the `BaselineDateType` enum and updating the query construction logic.

This functionality is useful for tracking the overall adoption of baseline features and understanding how the baseline status of features changes over time.

* rename