Remove the http middleware for caching
In this PR#1125, I moved the cache middleware to use the post request validation middleware hooks.
However, we depend on the status code to be set to know if we should cache the response. However,
the post requestion validation middleware hooks do not have access to the response code.
Or rather, it has not been set yet by the time the response body comes back to our cache middleware.

This means, we could either:
1. Move the cache middleware back to being a global cache layer but we have to pass data through other middlewares if we ever want to be careful about selective caching
2. Move the caching logic to the handlers([similar](https://github.com/search?q=repo%3AGoogleChrome%2Fchromium-dashboard%20rediscache&type=code) to ChromeStatus).

This change is the beginning of number 2 by removing the old code that won't be around afterwards.

Other changes:

- We still need the raw data cacher (provided by Valkey). We pass this into the http server setup so
  so that the handlers can use it to do type safe caching (parsing and caching).

Why is this change safe:
- Currently, the cache layer is not working. So we can do these adjustments.

This is part of a split up of https://github.com/GoogleChrome/webstatus.dev/compare/jcscottiii/fix-backend-cache?expand=1