Exclude discouraged features from spanner calls used on stats page (#1213)
This change modifies all of the spanner calls used on the stats page.

It adds a new getAllDiscouragedFeatureIDs method which is similar to
[getFeatureIDsForEachExcludedFeatureKey](https://github.com/GoogleChrome/webstatus.dev/blob/88b65fb0b5645ddc0b0b90ca46bfe9ab85e314db/lib/gcpspanner/excluded_feature_keys.go#L25-L52)

getFeatureIDsForEachExcludedFeatureKey and getAllDiscouragedFeatureIDs are both now called in a common method named getIgnoredFeatureIDsForStats that is used in each spanner query for the stats page.

This change comes from last week's feedback to ignore discouraged features.