Fix parsing of discouraged section
I forgot that we previously ignored the discouraged section when parsing
the raw data.json by setting it to nil. This change makes sure that we
include that data when parsing the file.

After this is merged, the web web_feature_consumer job will work and store
data in to the FeatureDiscouragedDetails table.

Other semi-related changes:
- Update the testdata file