Switch wpt metric to be `test_counts` view by default
This commit changes the default behavior for wpt_metric_view for:
- /v1/features
- /v1/features/{featureid}

This change will retrieve the WPT metrics as if view=test on wpt.fyi instead of the current behavior of view=subtest

On the frontend, allow the feature page (and chart panel) to configure which metric view to use by
reading the wpt_metric_view query parameter. We already do this on the overview page.

Add tests to assert the default behavior as well as behavior when explicitly setting the query parameter.