Adjust openapi for SavedSearchResponse
This change optionally returns bookmark status and role (if the user is authenticated) when looking at
a single or list of saved searches.

This data was already being returned by spanner but being discarded. This change
modifies the adapter layer to read that data and return the bookmark status and role.