Use K_REVISION for SPA bundle (#1270)
This should allow browsers to request a new bundle after a deployment

Cloud run environment provides [K_REVISION](https://cloud.google.com/run/docs/container-contract#:~:text=hello%2Dworld-,K_REVISION,-The%20name%20of)

So we can use that as an unique ID. We base 64 url encode it so that it is safe for browsers.

Also, set the local environment up with a fake value so that it replicates production.