Modify openapi & middleware to allow for optional auth (#1275)
Typically, you would set the security section to the following
to signal that endpoint can optionally take bearerAuth:

```
      security:
        - bearerAuth: []
        - {}
```

But the current Go openapi library does not let us know that {} has been set.

As a result, this change sets a new security called `noAuth`. Now with
this new security parameter, it lets the server know that:
1. If the authorization header is provided, try to authenticate like normal, or
2. If the authorization header is not set and this new noAuth option is a possibility
   for the route, treat it like an unauthenticated request that should proceed.

The GCIP middleware has been modified to fit that logic.

This is needed for the get_saved_search endpoint.

Other changes:
- Have the getSavedSearch endpoint use the new noAuth config
- Fix other openapi lint errors.