refactor: Add bookmarks context (#1307)
* refactor: Add bookmarks context

Currently, the code relies on the `DEFAULT_BOOKMARKS` constant to get the current list of bookmarks.

However, this leads to a problem where we can't have a dynamic list of bookmarks.

This is a refactoring that moves that information to a central location: webstatus-bookmarks-service.

Any child component can subscribe to that context to get updates.

Then in the future, we can have a process that can update the bookmark info.

Also beforehand, each component was figuring out the current bookmark. The service figures it out and publishes
that information too.

This refactor is part of the changes needed for Saved Searches (in which some bookmarks will be coming from the server.)

* address comments

* use helper one more place

* nit