Allow Updates of Browser Feature Availabilities (#1326)
Modified InsertBrowserFeatureAvailability to UpsertBrowserFeatureAvailability to support updates.

To support this, I added entityUniqueWriter to for this. Take a look at the constraint info section for more.

Database Constraint Note:

Due to the unique index on `(WebFeatureID, BrowserName)` in the
`BrowserFeatureAvailabilities` table, updates are handled via a
"delete then insert" pattern. The `entityUniqueWriter` is used to implement this behavior.

We were able to use most of the existing mapper functionality when using the entityUniqueWriter.
I got rid of the Merge method because we aren't really doing merges in this case. And I
also implemented the DeleteKey method to extract the keys to delete the row before inserting the new one.