Add user saved bookmark info to app-bookmark-info-context
Why: This is the pre-work for the changes to come to frontend/src/static/js/services/webstatus-bookmarks-service.ts for loading a saved search id from the query parameter.

This change adds two new fields to the context:
- userSavedSearchBookmarkTask: TaskTracker with information about the optional saved search bookmark being loaded when search_id query parameter is provided.
- currentLocation: A snapshot of the location information that applies to the info in the context. We will use this to let us know if our data is outdated.

This change also reorganizes the existing helpers into `bookmarkHelpers`. This will be useful for stubing. Because previously, we would run into this [issue](https://sinonjs.org/how-to/stub-dependency/)

There are two new helpers added:
- getCurrentQuery: A useful helper that lets the UI component know which query it should really use.
- isBusyLoadingBookmarkInfo: Let's the UI know that things aren't ready yet

Also, added some custom errors that we will be using in a future PR (this is a split up of https://github.com/GoogleChrome/webstatus.dev/compare/jcscottiii/ui-searches-v4.9?expand=1)
The errors provided a central place to get error messages for each type of error. In a future PR, we will show a toast for these messages. I added a catch all error `SavedSearchUnknownError`
that does print to console on purpose, thus the adjustment to frontend/web-test-runner.config.mjs to filter it our for tests

Other changes:
- Adjust existing code to use the bookmarkHelpers container
- Add api client call to getSavedSearchByID
- Add id to bookmark interface