Add task that automatically loads a search bookmark when search_id is present
Modified the WebstatusBookmarksService to have a new task loadingUserSavedBookmarkByIDTask.
This task is automatically triggered by changes to the location, api client and user (for sign in and sign out so that we can get role information too when calling getSavedSearchByID by providing a token)

Upon completion it updates _userSavedBookmarkByIDTaskTracker (which is used in refreshAppBookmarkInfo like the other internal properties).

If it errors, it will show a toast using the custom error messages from #1327 as well as update the page url to remove the faulty search_id

Other changes:
- Renamed changedLocation to _changedLocation to be consistent with the other internal properties
- Add pathname to AppLocation because it is needed by updatePageUrl