Add edit pencil icon to editable bookmarks in sidebar
If the user is the owner of the bookmark, they will now see the edit pencil
icon next to their bookmark.

When they click on the pencil, it will take them to the overview page with the query parameter `edit_bookmark` set to true.

In the future, there will be logic on the overview page to go straight into edit mode.