Refactor: Precise control over data fetching and task execution (#1334)
Background:

This PR refactors data fetching in two components by turning autoRun to false.
I originally observed an issue in Firefox in my original PR (https://github.com/GoogleChrome/webstatus.dev/pull/1330).
The firefox browser was seeing the tasks trigger multiple times and I thought my current handling of TaskNotReadyError and DuplicateErrors/TaskNotReadyError would handle it.
However, in lit task if a task triggers while there's a pending task already running, it would abort the first task. So as a result, sometimes firefox would
not get any information if the original task was aborted and the subsequent tasks were returning one of the DuplicateTaskErrors/TaskNotReadyError causing no more task executions to fully succeed.

This tries to simplify things by turning off autoRun and leveraging the willUpdate lit lifecycle method to precisely run a task. This allows us to not have to do checks on the task's arguments within the task itself since we control when the task exactly triggers.

Main changes:
- **`OverviewPage`:** The `loadingTask` is no longer set to `autoRun: true`. Instead, it is manually triggered in `willUpdate` based on changes to relevant properties and bookmark loading status. The `TaskNotReadyError` and related logic have been removed.
- **`webstatus-bookmarks-service`:**
    - The `loadingUserSavedBookmarkByIDTask` is no longer set to `autoRun: true` and is triggered in `willUpdate`.
    - The task now returns an object containing both the bookmark data and the current search query.

Other changes:
- **`webstatus-firebase-auth-service`:** Updated the type of the `user` state to `User | null | undefined`.
- Removed some redundant tests. Changed some tests to use sinon stubs
- Added some more coverage too.

This is a split up of #1330