Add adapter logic to update saved search (#1350)
* Add adapter logic to update saved search

Also updated the openapi document to follow AIP-134 for this PATCH method
https://google.aip.dev/134

This means adding an update_mask to the request.

This allows us to selectively update each field.

The adapter logic looks at the update_mask and converts it to the Optionally type in the gcpspanner layer

The validation of the request will happen in the http layer which will happen in the next PR.

* Update openapi/backend/openapi.yaml

Co-authored-by: Kyle Ju <kyleju@google.com>

* re-add min length to 1, add nullable for description

---------

Co-authored-by: Kyle Ju <kyleju@google.com>