feature: Add event handlers for saved search and bookmark actions (#1408)
This adds event listeners for the following events:
- saved-search-saved
- saved-search-edited
- saved-search-deleted
- saved-search-bookmarked
- saved-search-unbookmarked

Future PR(s) will emit those events. We create the event listeners so that anywhere in the app can emit an event and the appBookmarkInfo will update. And as a result, all the subscribers to appBookmarkInfo will get the new state automatically.

saved-search-bookmarked can use the same event listener method as saved-search-saved because:
1. The saved search owner can't toggle its bookmark status (like in  issue tracker).
2. As a result of point number 1, non owners can be the only users toggling. Which means it can use `handleSavedSearchSaved` since it just adds the saved search to the user's array locally.

The above also applies for saved-search-unbookmarked and saved-search-deleted

Other changes:
- On the overview page component, since we recently change the load task to disable autoRun and control when it runs, add a new check that verifies that we have not loaded this current saved search before.